requirements.txt
python-telegram-bot==21.6
gspread==6.1.2
google-auth==2.34.0
pydantic==2.9.2
loguru==0.7.2
python-dotenv==1.0.1

pytest==8.3.2
pytest-asyncio==0.23.8
pydantic-settings==2.4.0
